requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart==0.0.6
openai>=1.54.0
# numpy>=1.0,<2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.3.0
pinecone-client>=3.0.0
pinecone>=3.0.0
supabase>=2.10.0
python-docx==1.1.0
pypdf==3.17.4
pdfplumber==0.10.3
pillow>=10.3.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-dotenv==1.0.1
aiofiles==23.2.1
tiktoken>=0.8.0
openpyxl==3.1.2
unstructured==0.11.6
PyMuPDF>=1.23.0
pdf2image>=1.16.0
sendgrid>=6.11.0
